jsonc.json: Add Snippet "Project Tasks to diff and update a file".
Inserts a Task to diff a certain file of a project against its actually
used counterpart and update it in both directions.

Useful if you want to use Version Control for a config file and you
can't or don't want to create a Git repo in that file's parent dir
and/or edit the actually used file directly.